refactor: 온보딩 화면 코드 이동
