+++ app/src/main/java/com/practice/hanbitlunch/BlindarRoute.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavBackStackEntry
 import androidx.navigation.serialization.decodeArguments
 import com.practice.main.memo.MemoRoute
 import com.practice.main.nutrient.NutrientRoute
-import com.practice.onboarding.onboarding.OnboardingRoute
+import com.practice.onboarding.OnboardingRoute
 import com.practice.settings.feedback.FeedbackRoute
 import kotlinx.collections.immutable.toPersistentHashMap
 import kotlinx.serialization.InternalSerializationApi
